Update of constants for asset opcodes for new digital pairs
New pairs such as US-30 and others were added to IQ Option, and the script didn't have a method to update the asset opcodes for digital assets. Automatic opcode update for all pairs in binary, turbo, and digital has been added to the get_all_open_time() function. Now, just call this function to update the opcodes and use the charts for these new assets.